--- packages/giselle-engine/src/core/helpers/create-workspace.ts
@@ -1,6 +1,0 @@
-import { generateInitialWorkspace } from "@giselle-sdk/data-type";
-
-export async function createWorkspace() {
-	const workspace = generateInitialWorkspace();
-	return { ...workspace };
-}
