+++ packages/giselle-engine/src/core/workspaces/create-workspace.ts
@@ -1,6 +1,6 @@
 import { Workspace, generateInitialWorkspace } from "@giselle-sdk/data-type";
-import { setWorkspace } from "../helpers";
 import type { GiselleEngineContext } from "../types";
+import { setWorkspace } from "./utils";
 
 export async function createWorkspace(args: { context: GiselleEngineContext }) {
 	const workspace = generateInitialWorkspace();
