+++ packages/giselle-engine/src/core/workspaces/update-workspace.ts
@@ -1,6 +1,6 @@
 import type { Workspace, WorkspaceId } from "@giselle-sdk/data-type";
-import { setWorkspace } from "../helpers/set-workspace";
 import type { GiselleEngineContext } from "../types";
+import { setWorkspace } from "./utils";
 
 export async function updateWorkspace(args: {
 	context: GiselleEngineContext;
