+++ packages/giselle-engine/src/core/workspaces/utils.ts
@@ -1,6 +1,24 @@
 import { Workspace, type WorkspaceId } from "@giselle-sdk/data-type";
 import type { Storage } from "unstorage";
-import { workspacePath } from "./workspace-path";
+
+export function workspacePath(workspaceId: WorkspaceId) {
+	return `workspaces/${workspaceId}/workspace.json`;
+}
+
+export async function setWorkspace({
+	storage,
+	workspaceId,
+	workspace,
+}: {
+	storage: Storage;
+	workspaceId: WorkspaceId;
+	workspace: Workspace;
+}) {
+	await storage.setItem(workspacePath(workspaceId), workspace, {
+		// Disable caching by setting cacheControlMaxAge to 0 for Vercel Blob storage
+		cacheControlMaxAge: 0,
+	});
+}
 
 function parseAndRepairWorkspace(workspaceLike: unknown) {
 	const parseResult = Workspace.safeParse(workspaceLike);
